(read desc) On task4.py added phase portrait of lorenz attractor
On task4.py added the code to generate the phase portrait of the Lorenz attractor.
Of course task4.py has to be removed at a certain point, but for now I put the code there.
Start from that code for the second part of the task 4, where we have to plot a single trajectory (instead of the phase portrait obtained by running the system many times)